tests: fix tests
